Fix `clippy`
